add some verification against MATLAB output
Add a new test which checks that the output from the Python code matches
that of MATLAB to within machine precision.

We also need a nasty workaround for Python 2/3 incompatibility with npz.
The npz format uses pickle to store objects. The pickle format is
incompatible between PYthon 2 and 3 due to Unicode encoding issues. We
work around this by ensuring that we only store plain ol'arrays in the
npz file. It's mucky, but compatible.