add rotational symmetry mod, dtwavexfm2b and dtwaveifm2b
Originally based on PR#16 by @scf32. Pretty much a line-by-line copy of
dtwavexfm2b() as it appears in the MATLAB toolbox. You may prefer to
merge dtwavexfm2() and dtwavexfm2b() into one function which can handle
the extra coefficients returned by qshift and biort. As it is, the use
of 'qshift_b_bp' and 'near_sym_b_bp' (which I have included in .npz
format) is pretty much hardcoded in dtwavexfm2b.