do not copy results back from OpenCL by default
The new TransformDomainSignal representation for the output class has
allowed one to abstract the copying from device to host until the last
possible moment by introducing a compatible TransformDomainSignalOpenCl
class. The copy is performed at attribute access time. It may be
desirable to memoize the output which will be implemented in a separate
commit.