registration: refactor for speed
Thie is an API-breacking change. Instead of storing \tilde{Q} matrices
as arrays or array objects, move over to a multi-dimensional array
representation where the upper-triangular elements of the matrix are
stored directly in the last dimension.

This allows for a more efficient implementation of solvetransform()
which makes use of numpy's built in support for matrix stacks in
linalg.eigh.